Tweaked the conversion script. Using fenced colons for special style blocks.
Fenced colon start must separate with the content by a blank line.

Fenced colon start with class must be in two consecutive lines, but the class line need to be followed by a blank line before the content.

The content end and the ending three colons must separate by an empty line.